@@ -1,7 +1,9 @@
-package ca.cmpt276.restauranthealthinspection;
+package ui;
 
-import android.support.v7.app.AppCompatActivity;
+import androidx.appcompat.app.AppCompatActivity;
 import android.os.Bundle;
+
+import ca.cmpt276.restauranthealthinspection.R;
 
 public class MainActivity extends AppCompatActivity {
 
